pyproject.toml
[project]
name = "cel"
version = "0.1.0"
description = "CEL: Counterfactual Explanations Library"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Anonymous" },
]
maintainers = [
    { name = "Anonymous" },
]
keywords = [
    "counterfactual-explanations",
    "explainable-ai",
    "xai",
    "machine-learning",
    "normalizing-flows",
    "interpretability",
    "ppcef",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Typing :: Typed",
]

dependencies = [
    "torchdiffeq",
    "numpy",
    "scipy",
    "cvxpy",
    "scikit-learn",
    "hydra-core",
    "mlflow",
    "umap-learn",
    "cel-nflows>=0.14.1",
    "UMNN",
    "pandas",
    "torch",
    "tqdm",
    "matplotlib",
    "plotly",
    "dice-ml>=0.12",
    "mlxtend>=0.23.4",
    "pulp>=3.3.0",
    "lingam>=1.11.0",
    "networkx>=2.8,<3.0",
    "pytorch-tabnet",
    "pyfpgrowth",
    "scikit-image",
    "requests",
    "attrs",
    "dill",
    "transformers>=4.7.0,<5.0.0",
    "onnx<1.16.0",
    "ml-dtypes<0.5.0",
    "pyomo",
    "omlt",
    "tensorflow>=2.18.1",
    "alibi>=0.9.6",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pre-commit",
    "pytest",
    "ipykernel",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-material-extensions",
    "mkdocstrings[python]>=1.0.0",
    "mkdocs-gen-files>=0.6.0",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-section-index>=0.3.10",
]

[project.urls]
Homepage = "https://github.com/ofurman/counterfactuals"
Repository = "https://github.com/ofurman/counterfactuals"
Documentation = "https://github.com/ofurman/counterfactuals"
Issues = "https://github.com/ofurman/counterfactuals/issues"

[build-system]
requires = ["setuptools >= 77.0.3", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "ruff",
    "pre-commit",
    "pytest",
    "ipykernel",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["counterfactuals*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.ruff]
line-length = 100

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"counterfactuals/cf_methods/cet/*.py" = ["E741", "E711", "F841"]
"counterfactuals/cf_methods/ares/ares.py" = ["F841"]
"counterfactuals/cf_methods/globe_ce/globe_ce.py" = ["F841"]
"counterfactuals/pipelines/run_cet_pipeline.py" = ["F841"]
"counterfactuals/pipelines/run_lice_pipeline.py" = ["F841"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.13.3",
    "ipykernel>=6.29.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-material-extensions",
    "mkdocstrings[python]>=1.0.0",
    "mkdocs-gen-files>=0.6.0",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-section-index>=0.3.10",
]
